% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper-data.R
\docType{data}
\encoding{UTF-8}
\name{copper}
\alias{copper}
\alias{Cu}
\alias{bmk_worm}
\title{Biomarker Responses of the Ragworms to Copper and Warming}
\format{
An object of class \code{data.frame} with 60 rows and 6 columns.

An object of class \code{data.frame} with 210 rows and 22 columns.
}
\usage{
Cu

bmk_worm
}
\description{
Biomarker Responses of the Ragworms to Copper and Warming
}
\keyword{datasets}
