% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper_temp-data.R
\docType{data}
\encoding{UTF-8}
\name{copper_temp}
\alias{copper_temp}
\alias{Cu}
\alias{worm}
\title{Biomarker Responses of the Ragworms to Copper and Warming}
\format{
An object of class \code{data.frame} with 60 rows and 6 columns.

An object of class \code{data.frame} with 210 rows and 26 columns.
}
\usage{
Cu

worm
}
\description{
Biomarker Responses of the Ragworms to Copper and Warming
}
\keyword{datasets}
