% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctm-data.R
\docType{data}
\encoding{UTF-8}
\name{ctm}
\alias{ctm}
\alias{ctm_Cu}
\alias{ctm_worm}
\title{Biomarker Responses of the Ragworms to Copper and Warming}
\format{
An object of class \code{data.frame} with 60 rows and 7 columns.

An object of class \code{data.frame} with 210 rows and 26 columns.
}
\usage{
ctm_Cu

ctm_worm
}
\description{
Biomarker Responses of the Ragworms to Copper and Warming
}
\keyword{datasets}
