% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcalc.R
\name{diffcalc}
\alias{diffcalc}
\title{Calculate the Differences between Means}
\usage{
diffcalc(vec, control)
}
\arguments{
\item{vec}{a numeric vector, the mean responses.}

\item{control}{a logical, whether the control group exists.}
}
\description{
Calculate the Differences between Means
}
