% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensulizole-data.R
\docType{data}
\encoding{UTF-8}
\name{ensulizole}
\alias{ensulizole}
\alias{mussel_SoS}
\alias{mussel_gill}
\alias{mussel_digest}
\title{Biomarker Responses of the Blue Mussels to Organic UV Filters}
\format{
An object of class \code{data.frame} with 15 rows and 7 columns.

An object of class \code{data.frame} with 120 rows and 24 columns.

An object of class \code{data.frame} with 120 rows and 24 columns.
}
\source{
Pham, D. N., Sokolov, E. P., Falfushynska, H., & Sokolova, I. M.
  (2022). Gone with sunscreens: Responses of blue mussels (Mytilus edulis) to
  a wide concentration range of a UV filter ensulizole. Chemosphere, 309,
  136736. \doi{10.1016/j.chemosphere.2022.136736}.
}
\usage{
mussel_SoS

mussel_gill

mussel_digest
}
\description{
Biomarker Responses of the Blue Mussels to Organic UV Filters
}
\keyword{datasets}
