% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nolesser.R
\name{nolesser}
\alias{nolesser}
\title{Compare the Differences with Critical Values}
\usage{
nolesser(obs, cric)
}
\arguments{
\item{obs}{a numeric, the observed difference.}

\item{cric}{a numeric, the critical values of maximum absolute differences.}
}
\description{
Compare the Differences with Critical Values
}
