% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owlStat.R
\name{owlStat}
\alias{owlStat}
\title{Test Statistics for One-Way Layout Permutation Test}
\usage{
owlStat(lov, env = parent.frame())
}
\arguments{
\item{lov}{a list of vectors, responses by experimental groups.}

\item{env}{an environment, to access outer scope variables.}
}
\value{
\code{owlStat} returns a list with 5 components: \item{\code{n}}{the
  sample sizes.} \item{\code{avg}}{the mean responses.} \item{\code{T}}{the T
  statistic for global test.} \item{\code{d.multi}}{the differences in means
  for multiple comparisons.} \item{\code{mad}}{the maximum absolute
  differences in means.}.
}
\description{
\code{owlStat} computes statistics for \code{owl}. This is not meant to be
called directly.
}
\references{
Ernst, M. D. (2004). Permutation Methods: A Basis for Exact
  Inference. Statistical Science, 19(4), 676–685.
  \doi{10.1214/088342304000000396}.
}
