% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peramo-package.R
\docType{package}
\name{peramo-package}
\alias{peramo}
\alias{peramo-package}
\title{peramo: Permutation Tests for Randomization Model}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Perform permutation-based hypothesis testing for randomized experiments as suggested in Ludbrook & Dudley (1998) \doi{10.2307/2685470} and Ernst (2004) \doi{10.1214/088342304000000396}, introduced in Pham et al. (2022) \doi{10.1016/j.chemosphere.2022.136736}.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/phamdn/peramo/issues}
}

}
\author{
\strong{Maintainer}: Duy Nghia Pham \email{nghiapham@yandex.com} (\href{https://orcid.org/0000-0003-1349-1710}{ORCID})

Other contributors:
\itemize{
  \item Inna M. Sokolova (\href{https://orcid.org/0000-0002-2068-4302}{ORCID}) [thesis advisor]
}

}
\keyword{internal}
