% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twl.R
\name{twl}
\alias{twl}
\title{Two-Way Layout Permutation Test}
\usage{
twl(
  df,
  rand = 4999,
  seed = 1,
  mult = FALSE,
  type.mult = "control",
  alpha.mult = 0.05
)
}
\description{
Two-Way Layout Permutation Test
}
\examples{
manly2007 <- data.frame(
month = factor(rep(c("jun", "jul", "aug", "sep"), each = 6 ),
levels = c("jun", "jul", "aug", "sep")),
size = factor(rep(c("small", "large"), each = 3, times = 4),
levels = c("small", "large")),
consume = c( 13,242,105,182,21,7,8,59,20,24,312,68,515,488,88,460,1223,990,18,44,21,140,40,27))
twl(manly2007)

}
